requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.1
android==1.0
anyio==4.4.0
astroid==2.15.8
async-timeout==4.0.3
attrs==23.2.0
audiostream==0.2
Babel==2.9.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cmake @ file:///home/n0n3m4/Desktop/Pydroid/zBuildPrepared/cmake-3.26.4-py3-none-any.whl#sha256=695389dbf5c0aa220776821172e2f03865e0fe9ba10d7459373597d21af6643e
dill==0.3.7
distro==1.8.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isort==5.12.0
jedi==0.19.0
Kivy==2.2.1
lazy-object-proxy==0.0.0
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.0.5
ninja @ file:///home/n0n3m4/Desktop/Pydroid/zBuildPrepared/ninja-1.11.1-py3-none-any.whl#sha256=bde2ef34b4295ff07a0c1703ce5c7f3acccd621a91b21d4c6aaf09bd8f6591a8
openai==1.31.1
packaging==23.1
parso==0.8.3
pkgconfig==1.5.5
platformdirs==3.10.0
pyaes==1.6.1
pyasn1==0.6.0
pybind11==2.11.1
pycparser==2.21
pydantic==1.10.15
pygame==2.5.0
pyjnius==1.5.0
pylint==2.17.5
PySDL2==0.9.15
python-telegram-bot==21.2
pytz==2024.1
rsa==4.9
scikit-build==0.17.6
six==1.16.0
sniffio==1.3.1
telegram==0.0.1
Telethon==1.35.0
tomlkit==0.12.1
tqdm==4.66.4
typed-ast==1.5.5
typing_extensions==4.12.1
websocket-client==1.8.0
wrapt==1.15.0
yarl==1.9.4
